requirements.txt
git+https://github.com/ckan/ckanext-scheming.git#egg=ckanext-scheming
git+https://github.com/ckan/ckanext-dcat.git#egg=ckanext-dcat
rdflib==4.2.1
rdflib-jsonld==0.4.0
geomet>=0.2.0
ckantoolkit==0.0.3
future>=0.18.2
six
requests
cachetools